+++ tide/core/Options.h
@@ -72,6 +72,8 @@ class Options : public QObject, public boost::enable_shared_from_this<Options>
                 NOTIFY showTouchPointsChanged )
     Q_PROPERTY( bool showWindowBorders READ getShowWindowBorders
                 NOTIFY showWindowBordersChanged )
+    Q_PROPERTY( bool showWindowTitles READ getShowWindowTitles
+                WRITE setShowWindowTitles NOTIFY showWindowTitlesChanged )
     Q_PROPERTY( bool showZoomContext READ getShowZoomContext
                 NOTIFY showZoomContextChanged )
 
@@ -90,6 +92,7 @@ public:
     bool getShowTestPattern() const;
     bool getShowTouchPoints() const;
     bool getShowWindowBorders() const;
+    bool getShowWindowTitles() const;
     bool getShowZoomContext() const;
     //@}
 
@@ -97,6 +100,7 @@ public:
     //@{
     QColor getBackgroundColor() const;
     ContentPtr getBackgroundContent() const;
+    QString getBackgroundUri() const;
     //@}
 
     /**
@@ -121,6 +125,7 @@ public slots:
     void setShowTestPattern( bool set );
     void setShowTouchPoints( bool set );
     void setShowWindowBorders( bool set );
+    void setShowWindowTitles( bool set );
     void setShowZoomContext( bool set );
     //@}
 
@@ -135,6 +140,13 @@ public slots:
      *        A null pointer removes the current background.
      */
     void setBackgroundContent( ContentPtr content );
+
+    /**
+     * Set the background content from a uri.
+     * @param uri The uri of the content to set. If the uri is invalid, the
+     *        content is not modified. An empty uri removes the content.
+     */
+    void setBackgroundUri( const QString& uri );
     //@}
 
 signals:
@@ -144,11 +156,12 @@ signals:
     void autoFocusPixelStreamsChanged( bool set );
     void showContentTilesChanged( bool set );
     void showClockChanged( bool set );
-    void showControlAreaChanged();
+    void showControlAreaChanged( bool set );
     void showStatisticsChanged( bool set );
     void showTestPatternChanged( bool set );
     void showTouchPointsChanged( bool set );
     void showWindowBordersChanged( bool set );
+    void showWindowTitlesChanged( bool set );
     void showZoomContextChanged( bool set );
     //@}
 
@@ -172,6 +185,7 @@ private:
         ar & _showTouchPoints;
         ar & _showTestPattern;
         ar & _showWindowBorders;
+        ar & _showWindowTitles;
         ar & _showZoomContext;
         ar & _backgroundColor;
         ar & _backgroundContent;
@@ -186,6 +200,7 @@ private:
     bool _showTestPattern;
     bool _showTouchPoints;
     bool _showWindowBorders;
+    bool _showWindowTitles;
     bool _showZoomContext;
     QColor _backgroundColor;
     ContentPtr _backgroundContent;