+++ doc/Changelog.md
@@ -3,6 +3,8 @@ Changelog {#Changelog}
 
 # git master (1.1.0)
 
+* [34](https://github.com/BlueBrain/Tide/pull/34):
+  The options can be retrieved and modified through the REST interface.
 * [30](https://github.com/BlueBrain/Tide/pull/30):
   The contents can now be shown in fullscreen mode
 * [28](https://github.com/BlueBrain/Tide/pull/28):